Suggestion to fix player check before deleting box
The issue is, unless ALL players are near the box, the current version will delete the loot. With my changes, it won't delete the box if ANY player is nearby. 
May be more efficient to use nearEntities but I felt lazy and figured to fix it would at least be nice.